+++ cSploit/src/org/csploit/android/events/Os.java
@@ -4,11 +4,11 @@ package org.csploit.android.events;
  * os info found
  */
 public class Os implements Event {
-  public final short accuracy;
+  public final byte accuracy;
   public final String os;
   public final String type;
 
-  public Os(short accuracy, String os, String type) {
+  public Os(byte accuracy, String os, String type) {
     this.accuracy = accuracy;
     this.os = os;
     this.type = type;
